Move tests to Meta test class
